+++ examples/rest-example/src/main/java/com/accenture/examples/rest/HelloPoJo.java
@@ -53,7 +53,6 @@ public class HelloPoJo {
                         SamplePoJo.class.getName() + ", actual: " + response.getBody().getClass().getName());
             }
         }
-
     }
 
     @GET