+++ extensions/rest-automation/src/main/java/org/platformlambda/automation/services/ServiceResponseHandler.java
@@ -1,3 +1,21 @@
+/*
+
+    Copyright 2018-2020 Accenture Technology
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+ */
+
 package org.platformlambda.automation.services;
 
 import org.platformlambda.automation.config.RoutingEntry;
@@ -30,7 +48,6 @@ public class ServiceResponseHandler implements LambdaFunction {
     private static final Logger log = LoggerFactory.getLogger(ServiceResponseHandler.class);
 
     private static final SimpleXmlWriter xmlWriter = new SimpleXmlWriter();
-    private static ConcurrentMap<String, AsyncContextHolder> contexts;
 
     private static final String HEAD = "HEAD";
     private static final String STREAM = "stream";
@@ -42,6 +59,8 @@ public class ServiceResponseHandler implements LambdaFunction {
     private static final String HTML_END = "\n</pre>\n<body>\n</html>";
     private static final String RESULT = "result";
     private static final String ACCEPT_ANY = "*/*";
+
+    private ConcurrentMap<String, AsyncContextHolder> contexts;
 
     public ServiceResponseHandler(ConcurrentMap<String, AsyncContextHolder> contexts) {
         this.contexts = contexts;